[gtk] Non-panicking manual ApplicationBuilder impl
Closes gtk-rs/gtk4-rs#287
